Taxids with no NCBI match - continued
Removed the check and deletion of taxids with no NCBI match. This should only be necessary when working with a database combining sequences from NCBI with sequences from other databases. In that case, a dummy taxid may have been created for taxa not found in the NCBI taxonomy. Importantly, there was also a mistake in this part of the code, so contact the authors if you are using dummy taxids.