Update taxonomy.r
Added code needed when using a combined BOLD+nt database built with the MARES pipeline. Using this database may lead to taxids with no match in the NCBI taxonomy, so code to identify and handle this has also been added. In some cases, this situation can also occur when using the nt database, as it may not be completely synchronized with taxizedb. Added filtering of hits with low %identity and hits not identified to species level.